Move SceneManager to top level
Every other class, other than individual scenes, is top level.
Make SceneManager match this design.
